- Changed: Some experimental changes to reduce the memory footprint of Game objects.  For this first modification changed regen ticks, race, and speed from int to short.  These few changes reduce the Storm footprint by 2K.  All of these may be changed to ushort but at this time I'm not sure if any of the values can be negative.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@2410 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/ai/brain/StandardMobBrain.cs
@@ -152,7 +152,7 @@ namespace DOL.AI.Brain
             {
                 PathPoint path = MovementMgr.LoadPath(Body.PathID);
                 Body.CurrentWayPoint = path;
-                Body.MoveOnPath(path.MaxSpeed);
+				Body.MoveOnPath((short)path.MaxSpeed);
             }
 
             //If we are not attacking, and not casting, and not moving, and we aren't facing our spawn heading, we turn to the spawn heading
