+++ GameServer/realmabilities/handlers/IchorOfTheDeepAbility.cs
@@ -245,7 +245,7 @@ namespace DOL.GS.RealmAbilities
 			GameNPC npc = owner as GameNPC;
 			if (npc != null)
 			{
-				int maxSpeed = npc.MaxSpeed;
+				short maxSpeed = npc.MaxSpeed;
 				if (npc.CurrentSpeed > maxSpeed)
 					npc.CurrentSpeed = maxSpeed;
 			}
